refactor(native): extract sidecar archive validation
